Move pad casting test to test_shape_ops
